+++ packages/react/src/components/Checkbox/Checkbox.tsx
@@ -42,7 +42,7 @@ export const Checkbox = (props: CheckboxProps) => {
   return (
     <Flex className={clsx(componentClassName, className)} data-disabled={disabled ? '' : undefined}>
       <CheckboxPrimitive.Root
-        className="hearth-CheckboxRoot"
+        className={`${componentClassName}Root`}
         name={checkboxContext?.name}
         checked={checked}
         value={value}
@@ -64,7 +64,7 @@ export const Checkbox = (props: CheckboxProps) => {
           }
         }}
       >
-        <CheckboxPrimitive.Indicator asChild className="hearth-CheckboxIndicator">
+        <CheckboxPrimitive.Indicator asChild className={`${componentClassName}Indicator`}>
           <TickSmallIcon />
         </CheckboxPrimitive.Indicator>
       </CheckboxPrimitive.Root>