+++ packages/react/src/components/Container/Container.props.ts
@@ -1,5 +1,6 @@
 import { BorderColorProps } from '../../props/border-color.props';
-import { BorderProps } from '../../props/border.props';
+import { BorderStyleProps } from '../../props/border-style.props';
+import { BorderWidthProps } from '../../props/border-width.props';
 import { ColorProps } from '../../props/color.props';
 import { GridItemProps } from '../../props/grid-item.props';
 import { TextAlignProps } from '../../props/text-align.props';
@@ -11,7 +12,8 @@ export type ContainerProps = Omit<
   FlexProps,
   | 'as'
   | keyof ColorProps
-  | keyof BorderProps
+  | keyof BorderWidthProps
+  | keyof BorderStyleProps
   | keyof BorderColorProps
   | keyof GridItemProps
   | keyof TextTransformProps