+++ packages/react/src/components/Grid/Grid.tsx
@@ -16,11 +16,17 @@ import { flexItemPropDefs } from '../../props/flex-item.props';
 import { spacingPropDefs } from '../../props/spacing.props';
 import { backgroundColorPropDefs } from '../../props/background-color.props';
 import { positionPropDefs } from '../../props/position.props';
-import { borderPropDefs } from '../../props/border.props';
 import { textAlignPropDefs } from '../../props/text-align.props';
 import { textTransformPropDefs } from '../../props/text-transform.props';
 import { borderColorPropDefs } from '../../props/border-color.props';
 import { borderRadiusPropDefs } from '../../props/border-radius.props';
+import { alignItemsPropDefs } from '../../props/align-items.props';
+import { alignContentPropDefs } from '../../props/align-content.props';
+import { justifyContentPropDefs } from '../../props/justify-content.props';
+import { borderStylePropDefs } from '../../props/border-style.props';
+import { borderWidthPropDefs } from '../../props/border-width.props';
+import { zIndexPropDefs } from '../../props/z-index.props';
+import { overflowPropDefs } from '../../props/overflow.props';
 
 const COMPONENT_NAME = 'Grid';
 const componentClassName = withGlobalPrefix(COMPONENT_NAME);
@@ -36,6 +42,9 @@ export const Grid = (props: GridProps) => {
   } = extractProps(
     props,
     gridPropDefs,
+    alignItemsPropDefs,
+    alignContentPropDefs,
+    justifyContentPropDefs,
     positionPropDefs,
     paddingPropDefs,
     marginPropDefs,
@@ -43,14 +52,17 @@ export const Grid = (props: GridProps) => {
     backgroundColorPropDefs,
     gapPropDefs,
     sizePropDefs,
-    borderPropDefs,
+    borderStylePropDefs,
+    borderWidthPropDefs,
     borderRadiusPropDefs,
     borderColorPropDefs,
     gridItemPropDefs,
     flexItemPropDefs,
     spacingPropDefs,
     textAlignPropDefs,
-    textTransformPropDefs
+    textTransformPropDefs,
+    zIndexPropDefs,
+    overflowPropDefs
   );
 
   return (