+++ packages/react/src/props/background-color.props.ts
@@ -1,20 +1,14 @@
 import { CssTokenVariable } from '../types/css-token-variable';
 import { PropDef } from './prop-def';
 
-const colorValues = ['primary', 'secondary', 'brand'] as const;
+const values = ['primary', 'secondary', 'brand'] as const;
 
-const backgroundColorPropDefs = {
-  backgroundColor: { className: 'background-color', tokens: colorValues, responsive: false },
+export const backgroundColorPropDefs = {
+  backgroundColor: { className: 'bg', tokens: values, responsive: false },
 } satisfies {
-  backgroundColor: PropDef<(typeof colorValues)[number] | CssTokenVariable>;
+  backgroundColor: PropDef<(typeof values)[number] | CssTokenVariable>;
 };
 
-interface BackgroundColorProps {
-  /**
-   * Set the background colour.
-   */
-  backgroundColor?: (typeof colorValues)[number] | CssTokenVariable;
+export interface BackgroundColorProps {
+  backgroundColor?: (typeof values)[number] | CssTokenVariable;
 }
-
-export { backgroundColorPropDefs, colorValues as backgroundColorValues };
-export type { BackgroundColorProps };