+++ packages/react/src/props/border-radius.props.ts
@@ -1,50 +1,38 @@
 import { PropDef } from './prop-def';
 import { Responsive } from '../types/responsive';
 
-const borderRadiusValues = ['none', 'xs', 'sm', 'md', 'lg', 'xl', 'full', 'inherit'] as const;
+const values = ['none', 'xs', 'sm', 'md', 'lg', 'xl', 'full', 'inherit'] as const;
 
-const borderRadiusPropDefs = {
-  borderRadius: { className: 'border-radius', tokens: borderRadiusValues, responsive: true },
-  borderTopLeftRadius: {
-    className: 'border-top-left-radius',
-    tokens: borderRadiusValues,
-    responsive: true,
-  },
-  borderTopRightRadius: {
-    className: 'border-top-right-radius',
-    tokens: borderRadiusValues,
-    responsive: true,
-  },
-  borderBottomRightRadius: {
-    className: 'border-bottom-right-radius',
-    tokens: borderRadiusValues,
-    responsive: true,
-  },
-  borderBottomLeftRadius: {
-    className: 'border-bottom-left-radius',
-    tokens: borderRadiusValues,
-    responsive: true,
-  },
+export const borderRadiusPropDefs = {
+  borderRadius: { className: 'radius', tokens: values, responsive: true },
+  borderRadiusTopLeftNone: { className: 'radius-tl-none', responsive: false },
+  borderRadiusTopRightNone: { className: 'radius-tr-none', responsive: false },
+  borderRadiusBottomLeftNone: { className: 'radius-bl-none', responsive: false },
+  borderRadiusBottomRightNone: { className: 'radius-br-none', responsive: false },
+  borderRadiusTopNone: { className: 'radius-t-none', responsive: false },
+  borderRadiusRightNone: { className: 'radius-r-none', responsive: false },
+  borderRadiusBottomNone: { className: 'radius-b-none', responsive: false },
+  borderRadiusLeftNone: { className: 'radius-l-none', responsive: false },
 } satisfies {
-  borderRadius: PropDef<(typeof borderRadiusValues)[number]>;
-  borderTopLeftRadius: PropDef<(typeof borderRadiusValues)[number]>;
-  borderTopRightRadius: PropDef<(typeof borderRadiusValues)[number]>;
-  borderBottomRightRadius: PropDef<(typeof borderRadiusValues)[number]>;
-  borderBottomLeftRadius: PropDef<(typeof borderRadiusValues)[number]>;
+  borderRadius: PropDef<(typeof values)[number]>;
+  borderRadiusTopLeftNone: PropDef<boolean>;
+  borderRadiusTopRightNone: PropDef<boolean>;
+  borderRadiusBottomLeftNone: PropDef<boolean>;
+  borderRadiusBottomRightNone: PropDef<boolean>;
+  borderRadiusTopNone: PropDef<boolean>;
+  borderRadiusRightNone: PropDef<boolean>;
+  borderRadiusBottomNone: PropDef<boolean>;
+  borderRadiusLeftNone: PropDef<boolean>;
 };
 
-interface BorderRadiusProps {
-  /* Set the border radius. */
-  borderRadius?: Responsive<(typeof borderRadiusValues)[number]>;
-  /* Set the border-top-left radius. */
-  borderTopLeftRadius?: Responsive<(typeof borderRadiusValues)[number]>;
-  /* Set the border-top-right radius. */
-  borderTopRightRadius?: Responsive<(typeof borderRadiusValues)[number]>;
-  /* Set the border-bottom-right radius. */
-  borderBottomRightRadius?: Responsive<(typeof borderRadiusValues)[number]>;
-  /* Set the border-bottom-left radius. */
-  borderBottomLeftRadius?: Responsive<(typeof borderRadiusValues)[number]>;
+export interface BorderRadiusProps {
+  borderRadius?: Responsive<(typeof values)[number]>;
+  borderRadiusTopLeftNone?: boolean;
+  borderRadiusTopRightNone?: boolean;
+  borderRadiusBottomLeftNone?: boolean;
+  borderRadiusBottomRightNone?: boolean;
+  borderRadiusTopNone?: boolean;
+  borderRadiusRightNone?: boolean;
+  borderRadiusBottomNone?: boolean;
+  borderRadiusLeftNone?: boolean;
 }
-
-export { borderRadiusPropDefs };
-export type { BorderRadiusProps };