+++ packages/react/src/props/grid-item.props.ts
@@ -1,19 +1,19 @@
 import { Responsive, Union } from '../types/responsive';
 import { PropDef } from './prop-def';
 
-const columnsValues = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12'] as const;
+const values = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12'] as const;
 
-const gridItemPropDefs = {
-  gridColumnSpan: { className: 'grid-column-span', tokens: columnsValues, responsive: true },
-  gridArea: { className: 'grid-area', responsive: true },
-  gridColumn: { className: 'grid-column', responsive: true },
-  gridColumnStart: { className: 'grid-column-start', responsive: true },
-  gridColumnEnd: { className: 'grid-column-end', responsive: true },
-  gridRow: { className: 'grid-row', responsive: true },
-  gridRowStart: { className: 'grid-row-start', responsive: true },
-  gridRowEnd: { className: 'grid-row-end', responsive: true },
+export const gridItemPropDefs = {
+  gridColumnSpan: { className: 'grid-c-span', tokens: values, responsive: true },
+  gridArea: { className: 'grid-a', responsive: true },
+  gridColumn: { className: 'grid-c', responsive: true },
+  gridColumnStart: { className: 'grid-cs', responsive: true },
+  gridColumnEnd: { className: 'grid-ce', responsive: true },
+  gridRow: { className: 'grid-r', responsive: true },
+  gridRowStart: { className: 'grid-rs', responsive: true },
+  gridRowEnd: { className: 'grid-re', responsive: true },
 } satisfies {
-  gridColumnSpan: PropDef<(typeof columnsValues)[number]>;
+  gridColumnSpan: PropDef<(typeof values)[number]>;
   gridArea: PropDef<string>;
   gridColumn: PropDef<string>;
   gridColumnStart: PropDef<string>;
@@ -23,8 +23,8 @@ const gridItemPropDefs = {
   gridRowEnd: PropDef<string>;
 };
 
-interface GridItemProps {
-  gridColumnSpan?: Responsive<Union<string, (typeof columnsValues)[number]>>;
+export interface GridItemProps {
+  gridColumnSpan?: Responsive<Union<string, (typeof values)[number]>>;
   gridArea?: Responsive<string>;
   gridColumn?: Responsive<string>;
   gridColumnStart?: Responsive<string>;
@@ -33,6 +33,3 @@ interface GridItemProps {
   gridRowStart?: Responsive<string>;
   gridRowEnd?: Responsive<string>;
 }
-
-export { gridItemPropDefs };
-export type { GridItemProps };