+++ packages/react/src/props/margin.props.ts
@@ -3,35 +3,32 @@ import { CssTokenVariable } from '../types/css-token-variable';
 import { Responsive } from '../types/responsive';
 import { PropDef } from './prop-def';
 
-const marginValues = ['auto', ...spaceTokens] as const;
+const values = ['auto', ...spaceTokens] as const;
 
-const marginPropDefs = {
-  margin: { className: 'margin', tokens: marginValues, responsive: true },
-  marginTop: { className: 'margin-top', tokens: marginValues, responsive: true },
-  marginRight: { className: 'margin-right', tokens: marginValues, responsive: true },
-  marginBottom: { className: 'margin-bottom', tokens: marginValues, responsive: true },
-  marginLeft: { className: 'margin-left', tokens: marginValues, responsive: true },
-  marginInline: { className: 'margin-inline', tokens: marginValues, responsive: true },
-  marginBlock: { className: 'margin-block', tokens: marginValues, responsive: true },
+export const marginPropDefs = {
+  margin: { className: 'm', tokens: values, responsive: true },
+  marginTop: { className: 'mt', tokens: values, responsive: true },
+  marginRight: { className: 'mr', tokens: values, responsive: true },
+  marginBottom: { className: 'mb', tokens: values, responsive: true },
+  marginLeft: { className: 'ml', tokens: values, responsive: true },
+  marginX: { className: 'mx', tokens: values, responsive: true },
+  marginY: { className: 'my', tokens: values, responsive: true },
 } satisfies {
-  margin: PropDef<(typeof marginValues)[number] | CssTokenVariable>;
-  marginTop: PropDef<(typeof marginValues)[number] | CssTokenVariable>;
-  marginRight: PropDef<(typeof marginValues)[number] | CssTokenVariable>;
-  marginBottom: PropDef<(typeof marginValues)[number] | CssTokenVariable>;
-  marginLeft: PropDef<(typeof marginValues)[number] | CssTokenVariable>;
-  marginInline: PropDef<(typeof marginValues)[number] | CssTokenVariable>;
-  marginBlock: PropDef<(typeof marginValues)[number] | CssTokenVariable>;
+  margin: PropDef<(typeof values)[number] | CssTokenVariable>;
+  marginTop: PropDef<(typeof values)[number] | CssTokenVariable>;
+  marginRight: PropDef<(typeof values)[number] | CssTokenVariable>;
+  marginBottom: PropDef<(typeof values)[number] | CssTokenVariable>;
+  marginLeft: PropDef<(typeof values)[number] | CssTokenVariable>;
+  marginX: PropDef<(typeof values)[number] | CssTokenVariable>;
+  marginY: PropDef<(typeof values)[number] | CssTokenVariable>;
 };
 
-interface MarginProps {
-  margin?: Responsive<(typeof marginValues)[number] | CssTokenVariable>;
-  marginTop?: Responsive<(typeof marginValues)[number] | CssTokenVariable>;
-  marginRight?: Responsive<(typeof marginValues)[number] | CssTokenVariable>;
-  marginBottom?: Responsive<(typeof marginValues)[number] | CssTokenVariable>;
-  marginLeft?: Responsive<(typeof marginValues)[number] | CssTokenVariable>;
-  marginInline?: Responsive<(typeof marginValues)[number] | CssTokenVariable>;
-  marginBlock?: Responsive<(typeof marginValues)[number] | CssTokenVariable>;
+export interface MarginProps {
+  margin?: Responsive<(typeof values)[number] | CssTokenVariable>;
+  marginTop?: Responsive<(typeof values)[number] | CssTokenVariable>;
+  marginRight?: Responsive<(typeof values)[number] | CssTokenVariable>;
+  marginBottom?: Responsive<(typeof values)[number] | CssTokenVariable>;
+  marginLeft?: Responsive<(typeof values)[number] | CssTokenVariable>;
+  marginX?: Responsive<(typeof values)[number] | CssTokenVariable>;
+  marginY?: Responsive<(typeof values)[number] | CssTokenVariable>;
 }
-
-export { marginPropDefs };
-export type { MarginProps };